requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
matplotlib>=3.4.0
pillow>=8.0.0
tqdm>=4.65.0
scipy>=1.7.0
pytorch-fid>=0.3.0 
